requirements.txt
bottle==0.12.23
numpy==1.21.0
scipy==1.7.0
plotly==5.3.1
PyJWT==2.3.0

